refactor: Update page.tsx to use composition components
